split Matrix
