+++ src/matrix/internal/SliceColVector.ts
@@ -1,0 +1,120 @@
+/* *****************************************************************************
+ * Caleydo - Visualization for Molecular Biology - http://caleydo.org
+ * Copyright (c) The Caleydo Team. All rights reserved.
+ * Licensed under the new BSD license, available at http://caleydo.org/license
+ **************************************************************************** */
+/**
+ * Created by Samuel Gratzl on 04.08.2014.
+ */
+
+import {argSort, argFilter} from '../../index';
+import {list as rlist, RangeLike, Range, all, parse, Range1D} from '../../range';
+import {IValueType} from '../../datatype';
+import {IVector, IVectorDataDescription} from '../../vector';
+import AVector from '../../vector/AVector';
+import {IMatrix} from '../IMatrix';
+
+/**
+ * a simple projection of a matrix columns to a vector
+ */
+export default class SliceColVector extends AVector implements IVector {
+  readonly desc: IVectorDataDescription;
+  private colRange: Range1D;
+
+  constructor(private m: IMatrix, private col: number) {
+    super(null);
+    this.colRange = Range1D.from([this.col]);
+    this.desc = {
+      name: m.desc.name + '-c' + col,
+      fqname: m.desc.fqname + '-c' + col,
+      id: m.desc.id + '-c' + col,
+      type: 'vector',
+      idtype: m.rowtype,
+      size: this.dim[0],
+      value: this.valuetype,
+      description: m.desc.description,
+      creator: m.desc.creator,
+      ts: m.desc.ts
+    };
+    this.root = this;
+  }
+
+  persist() {
+    return {
+      root: this.m.persist(),
+      col: this.col
+    };
+  }
+
+  restore(persisted: any) {
+    let r: IVector = this;
+    if (persisted && persisted.range) { //some view onto it
+      r = r.view(parse(persisted.range));
+    }
+    return r;
+  }
+
+  get valuetype() {
+    return this.m.valuetype;
+  }
+
+  get idtype() {
+    return this.m.rowtype;
+  }
+
+  get idtypes() {
+    return [this.idtype];
+  }
+
+  size() {
+    return this.m.nrow;
+  }
+
+  /**
+   * return the associated ids of this vector
+   */
+  names(range?: Range): Promise<string[]> {
+    return this.m.rows(range);
+  }
+
+  ids(range?: RangeLike) {
+    return this.m.rowIds(range);
+  }
+
+  /**
+   * returns a promise for getting one cell
+   * @param i
+   */
+  at(i: number): Promise<IValueType> {
+    return this.m.at(i, this.col);
+  }
+
+  /**
+   * returns a promise for getting the data as two dimensional array
+   * @param range
+   */
+  data(range: RangeLike = all()): Promise<IValueType[]> {
+    const rr = parse(range);
+    const r = rlist(rr.dim(0), this.colRange);
+    return this.m.data(r).then((d) => {
+      if (d.length > 0 && Array.isArray(d[0])) {
+        return d.map((di) => di[0]);
+      }
+      return d;
+    });
+  }
+
+  sort(compareFn?: (a: IValueType, b: IValueType) => number, thisArg?: any): Promise<IVector> {
+    return this.data().then((d) => {
+      const indices = argSort(d, compareFn, thisArg);
+      return this.view(rlist(indices));
+    });
+  }
+
+  filter(callbackfn: (value: IValueType, index: number) => boolean, thisArg?: any): Promise<IVector> {
+    return this.data().then((d) => {
+      const indices = argFilter(d, callbackfn, thisArg);
+      return this.view(rlist(indices));
+    });
+  }
+}
